Added Graphical User Interface
PySlow is a Python-based tool for testing web servers against Slowloris attacks. Slowloris is a type of denial of service attack that aims to overwhelm a web server by opening and maintaining multiple connections to it, thereby exhausting its resources and preventing legitimate traffic from being served.

PySlow is a modified version of the original Slowloris tool developed by RSnake, with a graphical user interface (GUI) added by astrohippie. The tool is designed to be user-friendly and easy to use, even for those who may not have extensive experience with command-line tools or Python programming.

With PySlow, users can perform Slowloris attacks against target web servers, adjust various parameters, and monitor the progress of the attack in real-time through the GUI. PySlow is intended for use only in controlled environments for educational or testing purposes and should not be used for any malicious purposes.